requirements.txt
setuptools~=69.5
psycopg2~=2.9
xhtml2pdf~=0.2
whitenoise==6.5.0
Pillow==11.3.0
simplejson~=3.19
pytz==2023.3
django-widget-tweaks==1.4.12
django-environ==0.10.0
bradocs4py==1.3.1.0
boto3==1.34.69
gunicorn==21.2.0
django-allauth==0.54.0
Django==4.2
num2words
s3-streaming
django-storages==1.14.2
django-minify-html~=1.7